Move scope related files to proper folder
