+++ resharper/resharper-unity/src/CSharp/Feature/Services/LiveTemplates/Scope/UnityProjectScopeCategoryUIProvider.cs
@@ -1,12 +1,11 @@
 ﻿using System.Collections.Generic;
 using JetBrains.ReSharper.Feature.Services.LiveTemplates.Scope;
 using JetBrains.ReSharper.Feature.Services.LiveTemplates.Templates;
-using JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.LiveTemplates.Scope;
 using JetBrains.ReSharper.Plugins.Unity.Resources.Icons;
 using JetBrains.ReSharper.Psi.CSharp.Resources;
 using JetBrains.ReSharper.Psi.Resources;
 
-namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.LiveTemplates
+namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.LiveTemplates.Scope
 {
     // Defines a category for the UI, and the scope points that it includes
     [ScopeCategoryUIProvider(Priority = Priority, ScopeFilter = ScopeFilter.Project)]
