+++ resharper/resharper-unity/src/CSharp/Feature/Services/LiveTemplates/UnityTemplatesDefaultSettings.cs
@@ -4,7 +4,6 @@ using JetBrains.Application;
 using JetBrains.Application.Settings;
 using JetBrains.Diagnostics;
 using JetBrains.Lifetimes;
-using JetBrains.Util;
 
 namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.LiveTemplates
 {
